+++ src/main/java/com/raisetech/StudentManagement/controller/StudentController.java
@@ -1,0 +1,32 @@
+package com.raisetech.StudentManagement.controller;
+
+import com.raisetech.StudentManagement.data.Student;
+import com.raisetech.StudentManagement.data.StudentCourse;
+import com.raisetech.StudentManagement.service.StudentService;
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+
+public class StudentController {
+
+  private StudentService service;
+
+  @Autowired
+  public StudentController(StudentService service) {
+    this.service = service;
+  }
+
+  @GetMapping("/studentList")
+  public List<Student> getStudentList(){
+    return service.searchStudentList();
+  }
+
+  @GetMapping("/studentCourseList")
+  public List<StudentCourse> getStudentCourseList(){
+    return service.searchStudentCourseList();
+  }
+
+}
